pyproject.toml
[project]
name = "agent-share"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bikram-sambat>=0.1.1",
    "bs4>=0.0.2",
    "dateparser>=1.2.2",
    "dotenv>=0.9.9",
    "google-api-python-client>=2.179.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-genai>=1.30.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "langchain-mcp-adapters>=0.1.9",
    "langgraph>=0.6.5",
    "nepali-datetime>=1.0.8.4",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
